Split HIR expression typing helpers
Move numeric type predicates, method-owner typing, string borrow coercion, binary-add typing, and static expression value helpers out of the HIR monolith into hir/expressions.rs.

Update compiler source monolith reporting and ratchet ceilings for the new compiler.hir expression helper module.

Refs #1254.
