+++ scripts/ci/report-compiler-source-monoliths.py
@@ -28,6 +28,7 @@ BOUNDARY_MAP: dict[str, list[str]] = {
     "diagnostic_catalog.rs": ["compiler.diagnostics"],
     "diagnostics.rs": ["compiler.diagnostics"],
     "definitions.rs": ["compiler.hir"],
+    "expressions.rs": ["compiler.hir"],
     "generics.rs": ["compiler.hir"],
     "hir.rs": ["compiler.hir"],
     "json_contract.rs": ["compiler.commands"],
